### 2017-07-13 f9bc7f2
descriptive updates / authorship edits
