+++ README.Rmd
@@ -1,0 +1,123 @@
+---
+output: github_document
+---
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+```{r, echo = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "README-"
+)
+```
+
+# R/`origami`
+
+[![Travis-CI Build Status](https://travis-ci.org/jeremyrcoyle/origami.svg?branch=master)](https://travis-ci.org/jeremyrcoyle/origami)
+[![Build status](https://ci.appveyor.com/api/projects/status/i5qwp8cjb4j4x329?svg=true)](https://ci.appveyor.com/project/jeremyrcoyle/origami)
+[![Coverage Status](https://img.shields.io/codecov/c/github/jeremyrcoyle/origami/master.svg)](https://codecov.io/github/jeremyrcoyle/origami?branch=master)
+[![CRAN](http://www.r-pkg.org/badges/version/origami)](http://www.r-pkg.org/pkg/origami)
+[![Project Status: Active - The project has reached a stable, usable state and is being actively developed.](http://www.repostatus.org/badges/latest/active.svg)](http://www.repostatus.org/#active)
+[![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-blue.svg)](http://www.gnu.org/licenses/gpl-3.0)
+
+> High-powered framework for cross-validation: fold your data like it's paper!
+
+__Authors:__ [Jeremy Coyle](https://github.com/jeremyrcoyle) and [Nima
+Hejazi](http://nimahejazi.org)
+
+---
+
+## Description
+
+`origami` is an R package that provides a general framework for the application
+of cross-validation schemes to particular functions. By allowing arbitrary
+lists of results, `origami` accommodates a range of cross-validation applications.
+
+---
+
+## Installation
+
+For standard use, we recommend installing the package from
+[CRAN](https://cran.r-project.org/) via
+
+```{r cran-installation, eval = FALSE}
+install.packages("survtmle")
+```
+
+You can install a stable release of `origami` from GitHub via
+[`devtools`](https://www.rstudio.com/products/rpackages/devtools/) with:
+
+```{r gh-installation, eval = FALSE}
+devtools::install_github("jeremyrcoyle/origami")
+```
+
+---
+
+## Usage
+
+For details on how best to use `origami`, please consult the package vignette
+from within R, or browse the vignette
+[here](http://nimahejazi.org/origami/articles/generalizedCV.html).
+
+---
+
+## Example
+
+This minimal example shows how to use `origami` to apply cross-validation to the
+computation of a simple descriptive statistic using a sample data set. In
+particular, we obtain a cross-validated estimate of the mean:
+
+```{r example}
+data(mtcars)
+head(mtcars)
+library(origami)
+
+cvlm <- function(fold) {
+  train_data <- training(mtcars)
+  valid_data <- validation(mtcars)
+
+  mod <- lm(mpg ~ ., data = train_data)
+  preds <- predict(mod, newdata = valid_data)
+  list(coef = data.frame(t(coef(mod))), SE = ((preds - valid_data$mpg)^2))
+}
+
+folds <- make_folds(mtcars)
+results <- cross_validate(cvlm, folds)
+mean(results$SE)
+```
+
+---
+
+## Issues
+
+If you encounter any bugs or have any specific feature requests, please [file an
+issue](https://github.com/jeremyrcoyle/origami/issues).
+
+---
+
+## Citation
+
+After using the `origami` R package, please cite it:
+
+        @article{coyle2017origami,
+          doi = {},
+          url = {},
+          year  = {2017},
+          month = {},
+          publisher = {The Open Journal},
[…]
