+++ .travis.yml
@@ -5,7 +5,7 @@ warnings_are_errors: true
 
 r_packages:
   - covr
-  
+
 r:
   - oldrel
   - release
