### 2017-07-13 80c04db
add readme rmd, update docs
