+++ README.md
@@ -1,52 +1,97 @@
-# R/`origami`
 
-[![Travis-CI Build Status](https://travis-ci.org/jeremyrcoyle/origami.svg?branch=master)](https://travis-ci.org/jeremyrcoyle/origami)
-[![Build status](https://ci.appveyor.com/api/projects/status/i5qwp8cjb4j4x329?svg=true)](https://ci.appveyor.com/project/jeremyrcoyle/origami)
-[![Coverage Status](https://img.shields.io/codecov/c/github/jeremyrcoyle/origami/master.svg)](https://codecov.io/github/jeremyrcoyle/origami?branch=master)
-[![CRAN](http://www.r-pkg.org/badges/version/origami)](http://www.r-pkg.org/pkg/origami)
-[![Project Status: Active - The project has reached a stable, usable state and is being actively developed.](http://www.repostatus.org/badges/latest/active.svg)](http://www.repostatus.org/#active)
-[![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-blue.svg)](http://www.gnu.org/licenses/gpl-3.0)
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+R/`origami`
+===========
+
+[![Travis-CI Build Status](https://travis-ci.org/jeremyrcoyle/origami.svg?branch=master)](https://travis-ci.org/jeremyrcoyle/origami) [![Build status](https://ci.appveyor.com/api/projects/status/i5qwp8cjb4j4x329?svg=true)](https://ci.appveyor.com/project/jeremyrcoyle/origami) [![Coverage Status](https://img.shields.io/codecov/c/github/jeremyrcoyle/origami/master.svg)](https://codecov.io/github/jeremyrcoyle/origami?branch=master) [![CRAN](http://www.r-pkg.org/badges/version/origami)](http://www.r-pkg.org/pkg/origami) [![CRAN downloads](https://cranlogs.r-pkg.org/badges/origami)](https://CRAN.R-project.org/package=origami) [![Project Status: Active - The project has reached a stable, usable state and is being actively developed.](http://www.repostatus.org/badges/latest/active.svg)](http://www.repostatus.org/#active) [![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-blue.svg)](http://www.gnu.org/licenses/gpl-3.0)
 
 > High-powered framework for cross-validation: fold your data like it's paper!
 
----
+**Authors:** [Jeremy Coyle](https://github.com/jeremyrcoyle) and [Nima Hejazi](http://nimahejazi.org)
 
-## Description
+------------------------------------------------------------------------
 
-`origami` is an R package that provides a general framework for the application
-of cross-validation schemes to particular functions. By allowing arbitrary
-lists of results, `origami` accommodates a range of cross-validation applications.
+Description
+-----------
 
----
+`origami` is an R package that provides a general framework for the application of cross-validation schemes to particular functions. By allowing arbitrary lists of results, `origami` accommodates a range of cross-validation applications.
 
-## Installation
+------------------------------------------------------------------------
 
-The package will be available on [CRAN](https://cran.r-project.org/) in the
-near future; for now, download the latest version using
-[`devtools`](https://www.rstudio.com/products/rpackages/devtools/):
+Installation
+------------
 
+For standard use, we recommend installing the package from [CRAN](https://cran.r-project.org/) via
+
+``` r
+install.packages("survtmle")
 ```
+
+You can install a stable release of `origami` from GitHub via [`devtools`](https://www.rstudio.com/products/rpackages/devtools/) with:
+
+``` r
 devtools::install_github("jeremyrcoyle/origami")
 ```
 
----
+------------------------------------------------------------------------
 
-## Usage
+Usage
+-----
 
[…]
