### 2017-07-13 cf2407a
add parallel example
