+++ tests/testthat/test_future_plan.R
@@ -3,13 +3,11 @@ library(data.table)
 library(future)
 context("Future Plan")
 
-
 set.seed(1)
 
 data(mtcars)
 # make a lot of folds
 folds <- make_folds(mtcars, fold_fun = folds_bootstrap, V = 1000)
-
 
 # function to calculate cross-validated squared error
 cvlm <- function(fold) {
@@ -19,7 +17,6 @@ cvlm <- function(fold) {
     r <- lm(mpg ~ ., data = train_data)
     preds <- predict(r, newdata = valid_data)
     list(coef = data.frame(t(coef(r))), SE = ((preds - valid_data$mpg)^2))
-
 }
 
 plan(sequential)
@@ -33,9 +30,8 @@ time_mc <- system.time({
 
 if(future::availableCores() > 1) {
     test_that("MC is not significantly slower than sequential",{
-        #Windows doesn't support multicore
-        skip_on_os("windows")
-        expect_lt(time_mc["elapsed"], 1.2*time_seq["elapsed"])
+        skip_on_os("windows") # Windows doesn't support multicore
+        expect_lt(time_mc["elapsed"], 1.2 * time_seq["elapsed"])
       })
 }
 
