+++ R/cross_validate.R
@@ -21,7 +21,8 @@
 #'
 #' @export
 #'
-#' @example /inst/examples/cv_example.R
+#' @example /inst/examples/cv_simple_example.R
+#' @example /inst/examples/cv_parallel_example.R
 #'
 cross_validate <- function(cv_fun,
                            folds,
