### 2017-07-14 3a6dd6a
fix minor CRAN install mistake
