### 2017-07-16 64569a8
rm reliance on objects in the calling environment for example cv_funs; fix vignette refs
