+++ inst/examples/cv_parallel_example.R
@@ -1,34 +1,46 @@
+###############################################################################
 # This example explains how to use the cross_validate function with
 # parallelization using the framework of the future package.
+###############################################################################
 
-library(origami)
-library(data.table)
+suppressMessages(library(data.table))
 library(future)
-
+data(mtcars)
 set.seed(1)
 
-data(mtcars)
 # make a lot of folds
 folds <- make_folds(mtcars, fold_fun = folds_bootstrap, V = 1000)
 
-# function to calculate cross-validated squared error
-cvlm <- function(fold) {
-    train_data <- training(mtcars)
-    valid_data <- validation(mtcars)
+# function to calculate cross-validated squared error for linear regression
+cv_lm <- function(fold, data, reg_form) {
+  # get name and index of outcome variable from regression formula
+  out_var <- as.character(unlist(str_split(reg_form, " "))[1])
+  out_var_ind <- as.numeric(which(colnames(data) == out_var))
 
-    r <- lm(mpg ~ ., data = train_data)
-    preds <- predict(r, newdata = valid_data)
-    list(coef = data.frame(t(coef(r))), SE = ((preds - valid_data$mpg)^2))
+  # split up data into training and validation sets
+  train_data <- training(data)
+  valid_data <- validation(data)
+
+  # fit linear model on training set and predict on validation set
+  mod <- lm(as.formula(reg_form), data = train_data)
+  preds <- predict(mod, newdata = valid_data)
+
+  # capture results to be returned as output
+  out <- list(coef = data.frame(t(coef(mod))),
+              SE = ((preds - valid_data[, out_var_ind])^2))
+  return(out)
 }
 
 plan(sequential)
 time_seq <- system.time({
-    results_seq <- cross_validate(cvlm, folds)
+    results_seq <- cross_validate(cv_fun = cv_lm, folds = folds, data = mtcars,
+                                  reg_form = "mpg ~ .")
 })
 
 plan(multicore)
 time_mc <- system.time({
-    results_mc <- cross_validate(cvlm, folds)
+    results_mc <- cross_validate(cv_fun = cv_lm, folds = folds, data = mtcars,
+                                 reg_form = "mpg ~ .")
 })
 
 if(availableCores() > 1) {
