+++ README.Rmd
@@ -58,7 +58,7 @@ devtools::install_github("jeremyrcoyle/origami")
 ## Usage
 
 For details on how best to use `origami`, please consult the package vignette
-from within R, or browse the vignette
+from within [R](https://www.r-project.org/) or online
 [here](http://nimahejazi.org/origami/articles/generalizedCV.html).
 
 ---
@@ -70,21 +70,35 @@ computation of a simple descriptive statistic using a sample data set. In
 particular, we obtain a cross-validated estimate of the mean:
 
 ```{r example}
-data(mtcars)
-head(mtcars)
+set.seed(4795)
+library(stringr)
 library(origami)
 
-cvlm <- function(fold) {
-  train_data <- training(mtcars)
-  valid_data <- validation(mtcars)
+data(mtcars)
+head(mtcars)
 
-  mod <- lm(mpg ~ ., data = train_data)
+cv_lm <- function(fold, data, reg_form) {
+  # get name and index of outcome variable from regression formula
+  out_var <- as.character(unlist(str_split(reg_form, " "))[1])
+  out_var_ind <- as.numeric(which(colnames(data) == out_var))
+
+  # split up data into training and validation sets
+  train_data <- training(data)
+  valid_data <- validation(data)
+
+  # fit linear model on training set and predict on validation set
+  mod <- lm(as.formula(reg_form), data = train_data)
   preds <- predict(mod, newdata = valid_data)
-  list(coef = data.frame(t(coef(mod))), SE = ((preds - valid_data$mpg)^2))
+
+  # capture results to be returned as output
+  out <- list(coef = data.frame(t(coef(mod))),
+              SE = ((preds - valid_data[, out_var_ind])^2))
+  return(out)
 }
 
 folds <- make_folds(mtcars)
-results <- cross_validate(cvlm, folds)
+results <- cross_validate(cv_fun = cv_lm, folds = folds, data = mtcars,
+                          reg_form = "mpg ~ .")
 mean(results$SE)
 ```
 

[…]
