### 2017-07-17 97bb5a8
minor vignette edits, add startup message
