+++ R/zzz.R
@@ -1,0 +1,5 @@
+.onAttach <- function(...) {
+  packageStartupMessage("origami: Generalized Cross-Validation Framework")
+  packageStartupMessage("Version: ",
+                        utils::packageDescription("origami")$Version)
+}
