### 2017-07-17 0aa59b2
fix pandoc-citeproc 83
