+++ vignettes/refs.bib
@@ -1,86 +1,90 @@
-@article{VanDerLaan:2011ut,
-author = {Van Der Laan, M J and Rose, S},
-title = {{Targeted Learning: Causal Inference for Observational and Experimental Data - Mark J. van der Laan, Sherri Rose - Google Books}},
-year = {2011}
+@book{vdl2011targeted,
+  title={Targeted Learning: Causal Inference for Observational and Experimental
+    Data},
+  author={{van der Laan}, Mark J and Rose, Sherri},
+  year={2011},
+  publisher={Springer Science \& Business Media}
 }
 
-@article{vanderLaan:2007bz,
-author = {van der Laan, M J and Polley, E C and Hubbard, A E},
-title = {{Super learner}},
-journal = {Statistical Applications in Genetics and Molecular Biology},
-year = {2007},
-volume = {6},
-pages = {Art. 25--23 pp. (electronic)}
+@article{vdl2007super,
+  title={Super learner},
+  author={{van der Laan}, Mark J and Polley, Eric C and Hubbard, Alan E},
+  journal={Statistical applications in genetics and molecular biology},
+  volume={6},
+  number={1},
+  year={2007}
 }
 
-@article{Vaart:2006bz,
-author = {van der Vaart, A W and Dudoit, S and van der Laan, M J},
-title = {{Oracle inequalities for multi-fold cross validation}},
-journal = {Statistics {\&} Decisions},
-year = {2006},
-volume = {24},
-number = {3},
-pages = {351--371}
+@article{vaart2006oracle,
+  title={Oracle inequalities for multi-fold cross validation},
+  author={{van der Vaart}, Aad W and Dudoit, Sandrine and {van der Laan}, Mark
+    J},
+  journal={Statistics \& Decisions},
+  volume={24},
+  number={3},
+  pages={351--371},
+  year={2006}
 }
 
-%% Created using Papers on Tue, 21 Apr 2015.
-%% http://papersapp.com/papers/
-
-@article{Luedtke:2014uj,
-author = {Luedtke, A R and Van Der Laan, M J},
-title = {{Super-Learning of an Optimal Dynamic Treatment Rule}},
-year = {2014}
+@article{luedtke2016super,
+  title={Super-learning of an optimal dynamic treatment rule},
+  author={Luedtke, Alexander R and {van der Laan}, Mark J},
+  journal={The international journal of biostatistics},
+  volume={12},
+  number={1},
+  pages={305--332},
+  year={2016}
 }
 
-@article{VanDerLaan:2014tg,
-author = {Van Der Laan, M J and Luedtke, A R},
-title = {{Targeted learning of the mean outcome under an optimal dynamic treatment rule}},
-journal = {Journal of Causal Inference},
-year = {2014}
+@article{vdl2015targeted,
+  title={Targeted learning of the mean outcome under an optimal dynamic
+    treatment rule},
+  author={{van der Laan}, Mark J and Luedtke, Alexander R},
+  journal={Journal of causal inference},
+  volume={3},
+  number={1},
+  pages={61--95},
+  year={2015}
 }
 
-@article{VanDerLaan:2013tj,
-author = {Van Der Laan, M J and Hubbard, A E and Pajouh, S K},
-title = {{Statistical Inference for Data Adaptive Target Parameters}},
-year = {2013}
+@article{vdl2013statistical,
+  title={Statistical inference for data adaptive target parameters},
+  author={{van der Laan}, Mark J and Hubbard, Alan E and {Kherad-Pajouh}, Sara},
+  year={2013},
+  publisher={bepress}
 }
 
-@article{Zheng:2010ua,
-author = {Zheng, W and Van Der Laan, M J},
-title = {{Asymptotic theory for cross-validated targeted maximum likelihood estimation}},
[…]
