### 2017-07-17 0b57db7
update docs and links
