### 2017-09-24 #302 @nhejazi
Thanks very much for your efforts @karthik! No one in particular comes to mind - is there a list of JOSS reviewers we might be able to consult?

_Perhaps worth mentioning_ - this particular package provides a very general framework for easily implementing a variety of cross-validation schemes, making no domain expertise necessary to evaluate it. Really, its purpose is to make implementing cross-validation in a computationally efficient manner less of a headache. For now, we support some common CV schemes, and have the goal of adding more over time. Because of its generality, the important thing (in our eyes) in terms of reviewing the package is expertise in R and a working knowledge of cross-validation. I only mention the above in case it makes finding reviewers easier - hope this helps.
