### 2017-11-06 545af48
add folds2foldvec fun, updates docs, bumps verion
