+++ man/make_repeated_folds.Rd
@@ -18,5 +18,7 @@ and concatenating the results.
 }
 \seealso{
 Other fold generation functions: \code{\link{fold_from_foldvec}},
-  \code{\link{fold_funs}}, \code{\link{make_folds}}
+  \code{\link{fold_funs}}, \code{\link{folds2foldvec}},
+  \code{\link{make_folds}}
 }
+\concept{fold generation functions}
