+++ NAMESPACE
@@ -6,7 +6,9 @@ export(combiner_c)
 export(combiner_factor)
 export(combiner_rbind)
 export(cross_validate)
+export(fold_from_foldvec)
 export(fold_index)
+export(folds2foldvec)
 export(folds_bootstrap)
 export(folds_loo)
 export(folds_montecarlo)
@@ -19,6 +21,7 @@ export(make_folds)
 export(make_repeated_folds)
 export(training)
 export(validation)
+export(wrap_in_try)
 importFrom(abind,abind)
 importFrom(data.table,rbindlist)
 importFrom(future,future)
