### 2017-11-12 381043a
add hex sticker materials
