### 2017-11-12 2abf353
fix sticker path
