### 2017-11-12 9cecb8b
fix sticker type + position
