### 2017-11-12 29d1b4b
fix sticker position
