### 2017-12-17 #512 @whedon
[REVIEW]: origami - Cross-validation Framework
**Submitting author:** @jeremyrcoyle (<a href="http://orcid.org/0000-0002-9874-6649">Jeremy Coyle</a>)
**Repository:** <a href="https://github.com/jeremyrcoyle/origami" target ="_blank">https://github.com/jeremyrcoyle/origami</a>
**Version:** v0.8
