### 2017-12-17 #302 @arfon
@wrathematics - when you're ready, please head over to #512 for the actual review.
