### 2018-01-20 #512 @nhejazi
@wrathematics -- Thanks for submitting your review. The issue with the DOI pointing to the incorrect package version has been fixed by minting a new DOI via Zenodo (for the same GitHub release that we originally opened this review for). Here is a link to the new DOI: https://doi.org/10.5281/zenodo.1155901

Please let us know if there are any other corrections that we can make to complete this review process.
