### 2018-01-24 #512 @nhejazi
Perfect, thanks for confirming completion of the JOSS review.

Great to hear that you see the package as useful, and we're glad to hear the critiques/suggestions about the structure of the package, as we'd, of course, like to make sure that potential users feel the code to be natural. I've gone ahead and opened an issue with your comments [here](https://github.com/jeremyrcoyle/origami/issues/26) -- feel free to add to it at any time. I think it definitely makes sense to update the examples (`cv_lm` and the like) to match your suggestions, for example, with passing in a model formula rather than a string. In regard to a few of the other suggestions, I do think some of the ideas might clash with the design philosophy of `origami`, as the goal of the package is not to provide user-friendly functions like `cv_lm` but the rather the tools for interested users to build such functions (that is, users should really only be relying on functions like `training`, `validation`, and `cross_validate` to essentially write a cross-validation wrapper around practically any process). We'll likely have to think carefully about re-structuring aspects of the core functions to ensure that everything works smoothly with these refinements.
