### 2018-01-29 #512 @nhejazi
@karthik -- I believe we've now completed the review process, but I'm unsure of what the next steps might be in closing this review and generating the approved paper, etc. @wrathematics -- please feel free to correct me if I'm missing something here. Thank you all for your work in this review process.
