### 2018-01-30 #512 @nhejazi
Excellent -- we've archived v0.8.0 of the `origami` package and generated a DOI via Zenodo: DOI: 10.5281/zenodo.1155901. Please let me know if there's anything more we can do on our end.
