### 2018-01-30 #512 @arfon
@wrathematics many thanks for your review here and to @karthik for editing this submission.

@nhejazi @jeremyrcoyle - your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00512 ⚡️ 🚀 💥
